requirements.txt
mcp>=1.2.0
python-pptx>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
Pillow>=10.2.0
python-docx>=1.1.0
